Split the API into sub modules.
The new UI API was all located in one giant api.rs file. These have been
split up into multiple files.
